+++ lib/vagrant/action/vm/halt.rb
@@ -9,7 +9,7 @@ module Vagrant
 
         def call(env)
           if env[:vm].created? && env[:vm].vm.running?
-            env[:vm].system.halt if !env["force"]
+            env[:vm].guest.halt if !env["force"]
 
             if env[:vm].vm.state(true) != :powered_off
               env[:ui].info I18n.t("vagrant.actions.vm.halt.force")