+++ lib/vagrant/action/vm/host_name.rb
@@ -12,7 +12,7 @@ module Vagrant
           host_name = env[:vm].config.vm.host_name
           if !host_name.nil?
             env[:ui].info I18n.t("vagrant.actions.vm.host_name.setting")
-            env[:vm].system.change_host_name(host_name)
+            env[:vm].guest.change_host_name(host_name)
           end
         end
       end